requirements.txt
streamlit
pillow
python-dotenv
pybase64
transformers
diffusers
torch
accelerate